requirements.txt
-i https://pypi.org/simple
alembic==1.12.0; python_version >= '3.7'
aniso8601==9.0.1
attrs==23.1.0; python_version >= '3.7'
bcrypt==4.0.1; python_version >= '3.6'
blinker==1.6.2; python_version >= '3.7'
boto3==1.28.55; python_version >= '3.7'
botocore==1.31.55; python_version >= '3.7'
cachelib==0.10.2; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
flask==2.3.3; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.5; python_version >= '3.6'
flask-restplus==0.13.0
flask-script==2.0.6
flask-session==0.5.0; python_version >= '3.7'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flask-testing==0.8.1
greenlet==2.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonschema==4.19.1; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
packaging==23.1; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
psycopg2-binary==2.9.7; python_version >= '3.6'
pyjwt==2.8.0; python_version >= '3.7'
pytest==7.4.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
referencing==0.30.2; python_version >= '3.8'
rpds-py==0.10.3; python_version >= '3.8'
s3transfer==0.7.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.21; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.7; python_version >= '3.8'
